# Conflicts:
#	lichen/subsys/lichen/coap/coap_config.c
#	lichen/subsys/lichen/coap/coap_rangetest.c
#	lichen/subsys/lichen/coap/coap_status.c
#	lichen/subsys/lichen/coap/coap_waypoints.c
#	lichen/subsys/lichen/senml/senml.c
#	rust/lichen-gateway/src/tunnel_auth.rs

+++ lichen/subsys/lichen/coap/coap_slot_coord.c
@@ -264,7 +264,7 @@ bool lichen_slot_coord_validate_interleaved(const uint8_t *slots,
 	}
 
 	for (uint8_t i = 0; i < slot_count; i++) {
-		uint8_t expected = ordinal + i * gateway_count;
+		uint8_t expected = (uint8_t)(ordinal + i * gateway_count);
 		if (slots[i] != expected) {
 			return false;
 		}
